pyproject.toml
[project]
name = "robotics-stack"
version = "0.1.0"
description = "Modular interpretable robotics stack baseline."
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26.0",
  "pillow>=10.4.0",
]

[dependency-groups]
dev = [
  "mypy>=1.11.0",
  "pytest>=8.3.0",
  "ruff>=0.6.0",
]
sim = [
  "mujoco>=3.2.0",
  "opencv-python>=4.10.0",
  "pillow>=10.4.0",
]

[tool.pytest.ini_options]
addopts = "--basetemp=.pytest-tmp -p no:cacheprovider"
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
exclude = [
  ".pytest-cache",
  ".pytest-tmp",
  ".tmp-tests",
  "pytest-cache-files-*",
]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.mypy]
exclude = [
  "(^|\\\\)\\.pytest-cache($|\\\\)",
  "(^|\\\\)\\.pytest-tmp($|\\\\)",
  "(^|\\\\)\\.tmp-tests($|\\\\)",
  "(^|\\\\)pytest-cache-files-[^\\\\]+($|\\\\)",
]
cache_dir = ".mypy_cache"
python_version = "3.11"
strict = true
packages = ["apps", "interfaces", "modules"]

[[tool.mypy.overrides]]
module = ["cv2", "mujoco", "mujoco.*"]
ignore_missing_imports = true
